+++ lascli/parser/datasets.py
@@ -1,25 +1,28 @@
 import argparse
+import base64
 import csv
 import json
 import logging
 import yaml
 from argparse import ArgumentDefaultsHelpFormatter
 from collections import Counter
+from concurrent.futures import ThreadPoolExecutor, as_completed
+from functools import partial
+from itertools import zip_longest
 from pathlib import Path
 from time import time
-from itertools import zip_longest
-from functools import partial
+
+import filetype
 from las import Client
-from concurrent.futures import ThreadPoolExecutor
 
 from lascli.util import NotProvided, nullable, json_path
 
-# See https://docs.python.org/3/library/itertools.html
-def group(iterable, group_size, fillvalue=None):
-    "Collect data into fixed-length chunks or blocks"
-    # grouper('ABCDEFG', 4, 'x') --> ABCD EFGx"
+
+def group(iterable, group_size):
+    # See https://docs.python.org/3/library/itertools.html
+    # group('ABCDEFG', 4) --> ABCD EFG
     args = [iter(iterable)] * group_size
-    return zip_longest(*args, fillvalue=fillvalue)
+    return map(lambda g: list(filter(bool, g)), zip_longest(*args, fillvalue=None))
 
 
 def _create_documents_worker(t, client, dataset_id):
@@ -32,6 +35,31 @@ def _create_documents_worker(t, client, dataset_id):
         return doc, True, None
     except Exception as e:
         return doc, False, str(e)
+
+
+def _get_document_worker(las_client: Client, document_id, output_dir):
+    try:
+        document = las_client.get_document(document_id)
+        document_content = base64.b64decode(document['content'])
+        document_content_file_name = f'{document["documentId"]}.{filetype.guess_extension(document_content)}'
+        (output_dir / document_content_file_name).write_bytes(document_content)
+        document_ground_truth_file_name = f'{document["documentId"]}.json'
+        (output_dir / document_ground_truth_file_name).write_text(json.dumps(document.get('groundTruth', []), indent=2))
+        return document
+    except Exception as e:
+        print(f'Failed to download document {document_id}')
+        return None
+
+
+def _list_all_documents_in_dataset(las_client: Client, dataset_id):
+    list_response = las_client.list_documents(dataset_id=dataset_id)
+    yield from list_response['documents']
+    next_token = list_response.get('nextToken')
+
+    while next_token:
+        list_response = las_client.list_documents(dataset_id=dataset_id, next_token=next_token)
+        yield from list_response['documents']
+        next_token = list_response.get('nextToken')
 
 
 def post_datasets(las_client: Client, **optional_args):
@@ -125,7 +153,7 @@ def create_documents(
 
         for n, chunk in enumerate(group(documents, chunk_size)):
             fn = partial(_create_documents_worker, client=las_client, dataset_id=dataset_id)
-            inp = [(item, documents[item]) for item in chunk if item is not None and item not in uploaded_files]
+            inp = [(item, documents[item]) for item in chunk if item not in uploaded_files]
 
             for name, uploaded, reason in executor.map(fn, inp):
                 if uploaded:
@@ -139,12 +167,48 @@ def create_documents(
 
                 print(message)
 
-            minutes_spent = (time() - start_time) / 60
+            step_time = time() - start_time
+            minutes = int(step_time // 60)
+            seconds = int(step_time % 60)
             documents_processed = counter['uploaded'] + counter['failed']
             progress = documents_processed / num_docs * 100
-            print(f'{minutes_spent:.2f}m: {documents_processed}/{num_docs} docs processed | {progress:.1f}%')
+            print(f'{minutes}m{seconds}s: {documents_processed}/{num_docs} docs processed | {progress:.1f}%')
 
     return dict(counter)
+
+
+def get_documents(las_client: Client, dataset_id, output_dir, num_threads, chunk_size):
+    already_downloaded = set()
+    if output_dir.exists():
+        for path in output_dir.iterdir():
+            already_downloaded.add(path.stem)
+        print(f'Found {len(already_downloaded)} documents already downloaded')
+    else:
+        output_dir.mkdir()
+
+    not_downloaded = lambda document: document['documentId'] not in already_downloaded
+    print(f'Downloading documents in dataset {dataset_id} to {output_dir}')
+
+    with ThreadPoolExecutor(max_workers=num_threads) as executor:
+        start_time = time()
+        documents = filter(not_downloaded, _list_all_documents_in_dataset(las_client, dataset_id))
+        for chunk in group(documents, chunk_size):
+            futures = []
+            
+            for document in chunk:
+                futures.append(executor.submit(_get_document_worker, las_client, document['documentId'], output_dir))
+
+            for future in as_completed(futures):
+                document = future.result()
+                if document:
+                    already_downloaded.add(document['documentId'])
+
+            step_time = time() - start_time
+            minutes = int(step_time // 60)
+            seconds = int(step_time % 60)
+            print(f'{minutes}m{seconds}s: {len(already_downloaded)} downloaded')
+            
+    return {'Total downloaded documents': len(already_downloaded)}
 
 
 def create_datasets_parser(subparsers):
@@ -186,12 +250,12 @@ def create_datasets_parser(subparsers):
     delete_dataset_parser.add_argument('--delete-documents', action='store_true', default=False)
     delete_dataset_parser.set_defaults(cmd=delete_dataset)
 
-    upload_batch_to_dataset_parser = subparsers.add_parser(
+    create_documents_parser = subparsers.add_parser(
         'create-documents',
         formatter_class=argparse.RawTextHelpFormatter,
     )
-    upload_batch_to_dataset_parser.add_argument('dataset_id')
-    upload_batch_to_dataset_parser.add_argument(
+    create_documents_parser.add_argument('dataset_id')
+    create_documents_parser.add_argument(
         'input_path',
         default=False,
         type=Path,
@@ -203,28 +267,35 @@ def create_datasets_parser(subparsers):
              '3. Path to a csv file where each row contains information about one document, '
              'and the paths to the documents in the first column.'
     )
-    upload_batch_to_dataset_parser.add_argument('--chunk-size', default=500, type=int)
-    upload_batch_to_dataset_parser.add_argument(
+    create_documents_parser.add_argument('--chunk-size', default=500, type=int)
+    create_documents_parser.add_argument(
         '--documents-uploaded',
         default='.documents_uploaded.log',
         help='path to file used for caching progress (default: %(default)s)',
     )
-    upload_batch_to_dataset_parser.add_argument(
+    create_documents_parser.add_argument(
         '--documents-failed',
         default='.documents_failed.log',
         help='path to file used to store the documents that failed (default: %(default)s)',
     )
-    upload_batch_to_dataset_parser.add_argument('--num-threads', default=32, type=int, help='Number of threads to use')
-    upload_batch_to_dataset_parser.add_argument(
+    create_documents_parser.add_argument('--num-threads', default=32, type=int, help='Number of threads to use')
+    create_documents_parser.add_argument(
         '--ground-truth-encoding',
         default=None,
         help='Which encoding to use for the ground truth parsing',
     )
-    upload_batch_to_dataset_parser.add_argument(
+    create_documents_parser.add_argument(
         '--delimiter',
         default=',',
         help='delimiter to use if parsing a csv file',
     )
-    upload_batch_to_dataset_parser.set_defaults(cmd=create_documents)
+    create_documents_parser.set_defaults(cmd=create_documents)
+
+    get_documents_parser = subparsers.add_parser('get-documents')
+    get_documents_parser.add_argument('dataset_id')
+    get_documents_parser.add_argument('output_dir', type=Path, help='Path to download directory')
+    get_documents_parser.add_argument('--num-threads', default=32, type=int, help='Number of threads to use')
+    get_documents_parser.add_argument('--chunk-size', default=100, type=int)
+    get_documents_parser.set_defaults(cmd=get_documents)
 
     return parser