+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## Version 8.1.0 - 2022-03-29
+
+- Added `datasets get-documents`. This command will download all documents (PDF, JPEG etc. and JSON ground truth) to directory specified by `output_dir`
+- Show help when invoking `las` with no arguments
+- Support newer Python versions in setup.py
+- Added --version
+
 ## Version 8.0.0 - 2022-03-15
 
-- Fix bug in dataset create-documents
+- Fix bug in datasets create-documents
 - width and height is now optional in models create
 
 ## Version 7.1.0 - 2022-03-04